Finish renaming Procedures to Measurements
Full history of each file can be tracked using git log --follow, or GitHub's blame view.
