+++ .gitignore
@@ -6,7 +6,7 @@ Utilities/__pycache__
 Utilities/test.csv
 Utilities/.test.txt
 *.pyc
-Procedures/__pycache__/planefit.cpython-35.pyc
+Measurements/__pycache__/planefit.cpython-35.pyc
 *.pyc
 
 # Squid array amplifier settings #
