+++ Utilities/setup/DESKTOP-KDO0NTE.py
@@ -20,20 +20,20 @@ from Nowack_Lab.Utilities import utilities, save, logging, conversions, anim, co
 from Nowack_Lab.Utilities.plotting import plot_mpl
 from Nowack_Lab.Utilities.save import Measurement
 from Nowack_Lab.Utilities.say import say
-from Nowack_Lab.Procedures.daqspectrum import DaqSpectrum, SQUIDSpectrum
-from Nowack_Lab.Procedures.dctransport import DAQ_IV
-from Nowack_Lab.Procedures.heightsweep import Heightsweep
-from Nowack_Lab.Procedures.mod2D import Mod2D
-from Nowack_Lab.Procedures.mutual_inductance import MutualInductance
-from Nowack_Lab.Procedures.navigation import move
-from Nowack_Lab.Procedures.planefit import Planefit
-from Nowack_Lab.Procedures.scanline import Scanline
-from Nowack_Lab.Procedures.scanplane import Scanplane
-from Nowack_Lab.Procedures.scanspectra import Scanspectra
-from Nowack_Lab.Procedures.squidIV import SquidIV
-from Nowack_Lab.Procedures.touchdown import Touchdown
-from Nowack_Lab.Procedures.transport import RvsVg, RvsTime, RvsT, RvsVg_Vtg, RvsT_RT_to_4K, RvsVg_T
-from Nowack_Lab.Procedures.magnetotransport import RvsB, RvsVg_B, RvsB_BlueFors, RvsVg_B_BlueFors
+from Nowack_Lab.Measurements.daqspectrum import DaqSpectrum, SQUIDSpectrum
+from Nowack_Lab.Measurements.dctransport import DAQ_IV
+from Nowack_Lab.Measurements.heightsweep import Heightsweep
+from Nowack_Lab.Measurements.mod2D import Mod2D
+from Nowack_Lab.Measurements.mutual_inductance import MutualInductance
+from Nowack_Lab.Measurements.navigation import move
+from Nowack_Lab.Measurements.planefit import Planefit
+from Nowack_Lab.Measurements.scanline import Scanline
+from Nowack_Lab.Measurements.scanplane import Scanplane
+from Nowack_Lab.Measurements.scanspectra import Scanspectra
+from Nowack_Lab.Measurements.squidIV import SquidIV
+from Nowack_Lab.Measurements.touchdown import Touchdown
+from Nowack_Lab.Measurements.transport import RvsVg, RvsTime, RvsT, RvsVg_Vtg, RvsT_RT_to_4K, RvsVg_T
+from Nowack_Lab.Measurements.magnetotransport import RvsB, RvsVg_B, RvsB_BlueFors, RvsVg_B_BlueFors
 from Nowack_Lab.Instruments.attocube import Attocube
 from Nowack_Lab.Instruments.keithley import Keithley2400, Keithley2450, Keithley2600, KeithleyPPMS
 from Nowack_Lab.Instruments.lakeshore import Lakeshore372
